Fix edgecases around 12/24h boundaries with streaks
There were two main things, here. There was a bug where
if medication was taken within 24 hours of each other, the streak would
not count as two days, even if they were literally on two different
days. While doing the calculation based on dates this worked in isolation,
it broke the case where taking the medication the next day (but more
than 24h) keeps the streak (it would add an extra day due to the extra
time having elapsed). It was at this time I realized that the use of
JULIANDAY calculated the time based on it being noontime, and the fact
we were using casts as a floor intoroduced disparities. I changed this
to a conversion based on unix time (which starts at midnight) to correct
this disparity.